Review: move Rankings selector fn to separatte utils file
